Fix fractional zoom levels in the legacy api

+++ app/controllers/legacy_image_service_controller.rb
@@ -72,7 +72,7 @@ class LegacyImageServiceController < ApplicationController
     case
     when params[:region] && params[:zoom]
       x, y, w, h = params[:region].split(',').map(&:to_i)
-      zoom = params[:zoom].to_i / 100.0
+      zoom = params[:zoom].to_f / 100.0
       [x / zoom, y / zoom, w / zoom, h / zoom].map(&:to_i).join(',')
     when params[:region]
       params[:region]
